Updates to Outputs for peggy parsed data Bug fix for zero not registering as a valid number throughout the forms
