+++ Emrald-UI/src/components/forms/ActionForm/FormFieldsByType/RunApplication/CustomForms/MAAP/FormFieldsByType/Outputs.tsx
@@ -1,4 +1,5 @@
 import {
+  Autocomplete,
   Box,
   Divider,
   FormControl,
@@ -6,6 +7,7 @@ import {
   InputLabel,
   MenuItem,
   Select,
+  TextField,
   Typography,
 } from '@mui/material';
 import { useVariableContext } from '../../../../../../../../contexts/VariableContext';
@@ -14,18 +16,13 @@ import { useEffect, useState } from 'react';
 import { useCustomForm } from '../../useCustomForm';
 
 const Outputs = () => {
-  const [docLinkVariable, setDocLinkVariable] = useState<string>('');
-  const [output, setOutput] = useState<string>('');
-  const { variableList } = useVariableContext();
   const { formData, setFormData } = useCustomForm();
+  const [docLinkVariable, setDocLinkVariable] = useState<string>(formData.docLinkVariable || '');
+  const [output, setOutput] = useState<string>(formData.output || '');
+  const { variableList } = useVariableContext();
   const docLinkVariables = variableList.value
-    .filter(({ varScope, docType }) => varScope === 'gtDocLink' && docType === 'dtXML')
+    .filter(({ varScope }) => varScope === 'gtDocLink')
     .map(({ name }) => name);
-
-  useEffect(() => {
-    setDocLinkVariable(formData.docLinkVariable || '');
-    setOutput(formData.output || '');
-  }, []);
 
   useEffect(() => {
     setFormData((prevFormData: any) => ({
@@ -43,29 +40,15 @@ const Outputs = () => {
 
       <Box>
         <Box display={'flex'} alignItems={'center'}>
-          <FormControl sx={{ m: 1, minWidth: 200 }} size="small">
-            <InputLabel id="doc-link-variables-label">Doc Link Variables</InputLabel>
-            <Select
-              labelId="doc-link-variables-label"
-              id="doc-link-variables"
-              label="Doc Link Variables"
-              value={docLinkVariable}
-              onChange={(event) => setDocLinkVariable(event.target.value)}
-              MenuProps={{
-                PaperProps: {
-                  style: {
-                    maxHeight: 350,
-                  },
-                },
-              }}
-            >
-              {docLinkVariables.map((variable, index) => (
-                <MenuItem key={index} value={variable}>
-                  {variable}
-                </MenuItem>
-              ))}
-            </Select>
-          </FormControl>
+          <Autocomplete
+            freeSolo
+            options={docLinkVariables}
+            defaultValue={formData.docLinkVariable || ''}
+            renderInput={(params) => <TextField {...params} label="Doc Link Variables" />}
+            onChange={(_, event) => setDocLinkVariable(event || '')}
+            sx={{ width: '250px' }}
+            size="small"
+          />
           <FormControl sx={{ m: 1, minWidth: 200 }} size="small">
             <InputLabel id="output-label">Output</InputLabel>
             <Select
