+++ Emrald-UI/src/components/forms/ActionForm/FormFieldsByType/RunApplication/CustomForms/MAAP/FormFieldsByType/Parameters.tsx
@@ -8,28 +8,19 @@ import {
   Checkbox,
   MenuItem,
 } from '@mui/material';
-import { Parameter, Value } from '../maap';
 import { useEffect, useState } from 'react';
 import { SelectComponent } from '../../../../../../../common';
 import { appData } from '../../../../../../../../hooks/useAppData';
 import { useActionFormContext } from '../../../../../ActionFormContext';
-
-export const getParameterName = (row: Parameter) => {
-  let name =
-    row.target.type === 'call_expression'
-      ? ((row.target.value as Value).value as string)
-      : (row.target.value as string);
-  if (row.target?.arguments && row.target.arguments.length > 0) {
-    name = name + `(${String(row.target.arguments[0].value)})`;
-  }
-  return name;
-};
+import { Parameter } from '../../../CustomApplicationTypes';
+import useRunApplication from '../../../useRunApplication';
 
 const Parameters = () => {
   const { formData, setFormData } = useActionFormContext();
   const [useVariable, setUseVariable] = useState<{ [key: string]: boolean }>({});
   const [variable, setVariable] = useState<{ [key: string]: string }>({});
   const [parameters, setParameters] = useState<Parameter[]>([]);
+  const { getParameterName } = useRunApplication();
 
   useEffect(() => {
     setUseVariable(
