+++ Emrald-UI/src/components/layout/Header/menuOptions.ts
@@ -20,7 +20,7 @@ export const projectOptions: MenuOption[] = [
       // Create a new file input element
       const fileInput = document.createElement('input');
       fileInput.type = 'file'; // Set input type to file
-      fileInput.accept = '.json'; // Specify accepted file types as JSON
+      fileInput.accept = '.json, .emrald'; // Specify accepted file types as JSON
       fileInput.style.display = 'none'; // Hide the file input element
 
       // Function to handle file selection
@@ -38,7 +38,7 @@ export const projectOptions: MenuOption[] = [
           try {
             const parsedContent = JSON.parse(content);
               if (parsedContent && parsedContent.hasOwnProperty('emraldVersion')) {
-                populateNewData(content);
+                populateNewData(parsedContent);
               } else {
                 const upgradedModel = upgradeModel(content);
                 if (upgradedModel) {
